+++ index.js
@@ -7,3 +7,4 @@ require('./lib/validations.js');
 require('./lib/cookies.js');
 require('./lib/localSessionStorage.js');
 require('./lib/mobile.js');
+require('./lib/execute.js');