+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Current
 - :rocket: updated wdio version to 8
 - :rocket: added I press Y key X times step
+- :rocket: added I execute steps allow to execute client functions
 
 ## 0.0.13
 - :rocket: added I wait until current url step